refactor: ResponseTemplate common 모듈로 이동
