+++ common/src/main/java/itcast/ResponseTemplate.java
@@ -1,4 +1,4 @@
-package itcast.dto.response;
+package itcast;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
